backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Caching & Task Queue
redis==5.0.1
celery==5.3.6

# HTTP & Scraping
httpx==0.26.0
beautifulsoup4==4.12.3
feedparser==6.0.10
lxml==5.1.0

# AI & ML
openai==1.10.0
tiktoken==0.5.2

# Vector Databases
chromadb==0.4.22
pinecone-client==3.0.0

# Scheduling
apscheduler==3.10.4

# Utils
python-dotenv==1.0.0
python-multipart==0.0.6

# Dev & Testing
pytest==7.4.4
pytest-asyncio==0.23.3